requirements.txt
Django==4.2
mongoengine==0.29.1
django-ninja==0.22.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cloudinary==1.29.0
opencv-python==4.8.0.76
pyyaml==6.0.3
natsort==8.2.0
pydantic==1.10.12
uvicorn==0.22.0
fastapi==0.95.2
